MyMiniStore/my_shop/requirements.txt
# ===============================================
# ОСНОВНІ ЗАЛЕЖНОСТІ ПРОЄКТУ (Django 5.x)
# ===============================================

# Фреймворк Django
Django~=5.0.0 

# Сервер WSGI для продакшену (використовується в Docker)
gunicorn

# ===============================================
# СТАТИЧНІ ФАЙЛИ та МЕДІА (для продакшену)
# ===============================================

# Для коректного обслуговування статичних файлів у продакшені (особливо важливо при використанні Nginx/Gunicorn)
whitenoise

# ===============================================
# БАЗА ДАНИХ
# ===============================================

# Адаптер для SQLite3 входить до Python, тут не потрібен.
# Якщо ви використовували PostgreSQL:
# psycopg2-binary

# ===============================================
# ДОДАТКОВІ БІБЛІОТЕКИ (Згідно з вашим settings.py)
# ===============================================

# Django Rest Framework (для API)
djangorestframework

# Django Filters (для фільтрації даних в API, вирішення проблеми six)
# Вимагає версії 23.2 або новішої для сумісності з Django 4.x/5.x
django-filter>=23.2
python-dotenv
psycopg2-binary
pillow>=12.0.0
# ===============================================
# ІНТЕГРАЦІЇ ТА ФУНКЦІОНАЛ (Згідно з вашим views.py)
# ===============================================

# Бібліотека для взаємодії з LiqPay (припускаю, що ви її використовуєте для payment)
# Це умовна назва, залежно від того, яку бібліотеку ви обрали
# Наприклад:
# requests 
# PyOpenSSL # Часто потрібен для LiqPay
# cryptography # Часто потрібен для безпеки